CI-73 Updated indexer to use Elasticsearch 5

+++ src/Sfa.Das.Sas.Infrastructure/Shared/Elasticsearch/ElasticsearchIndexMaintainerBase.cs
@@ -99,7 +99,7 @@
             foreach (var bulkResponse in elementIndexResult)
             {
                 totalCount += bulkResponse.Items.Count();
-                took += bulkResponse.Took;
+                took += (int)bulkResponse.Took;
                 errorCount += bulkResponse.ItemsWithErrors.Count();
             }
